+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.application;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.event.MenuCreatedEvent;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menu.ui.exception.NoMenuGroupException;
+import kitchenpos.menu.exception.NoMenuGroupException;
 import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.dto.MenuRequest;
