+++ src/main/java/kitchenpos/menu/exception/IllegalMenuPriceException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui.exception;
+package kitchenpos.menu.exception;
 
 public class IllegalMenuPriceException extends RuntimeException{
     public IllegalMenuPriceException(String message) {
