+++ src/main/java/kitchenpos/menu/exception/NoMenuGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui.exception;
+package kitchenpos.menu.exception;
 
 public class NoMenuGroupException extends RuntimeException{
     public NoMenuGroupException() {
