+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.menu.ui;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menu.ui.exception.IllegalMenuPriceException;
-import kitchenpos.menu.ui.exception.NoMenuGroupException;
+import kitchenpos.menu.exception.IllegalMenuPriceException;
+import kitchenpos.menu.exception.NoMenuGroupException;
 import kitchenpos.menuproduct.exception.NoMenuProductException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
