+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.ui.exception.IllegalMenuPriceException;
-import kitchenpos.menu.ui.exception.NoMenuGroupException;
-import kitchenpos.menuproduct.exception.NoMenuProductException;
+import kitchenpos.menu.exception.IllegalMenuPriceException;
+import kitchenpos.menu.exception.NoMenuGroupException;
 import kitchenpos.menuproduct.domain.MenuProductRepository;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menugroup.domain.MenuGroup;
@@ -26,7 +25,6 @@ import org.springframework.context.ApplicationEventPublisher;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
