+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.domain;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.ui.exception.IllegalMenuPriceException;
+import kitchenpos.menu.exception.IllegalMenuPriceException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
